Fix type errors in decoder when input data is not double.

+++ +JPEG/decoder.m
@@ -198,7 +198,7 @@ classdef decoder < handle
                 quantisedChannel{c} = cell2mat(reshape(blocksWithACCoefficientsReordered{c}, obj.componentSizeInBlocks{c}(1), obj.componentSizeInBlocks{c}(2)).');
 
                 % Dequantise
-                dequantisedChannel{c} = blkproc(quantisedChannel{c}, [8 8], ...
+                dequantisedChannel{c} = blkproc(double(quantisedChannel{c}), [8 8], ...
                                                 @(block)(block.*obj.quantisationTables{obj.quantisationTableDestinationSelector(channelID) + 1}));
 
                 % IDCT
@@ -318,7 +318,7 @@ classdef decoder < handle
 
         % TODO: MAKE THESE INTO UTILITIES!
         function short = getNumericShort(obj, startByte)
-            short = (obj.inputStruct.numericData(startByte) * 256) + obj.inputStruct.numericData(startByte + 1);
+            short = (uint16(obj.inputStruct.numericData(startByte)) * 256) + uint16(obj.inputStruct.numericData(startByte + 1));
         end
 
         function [high low] = getNumericNibblesFromByte(obj, startByte)
@@ -333,7 +333,7 @@ classdef decoder < handle
             %
 
             % Lf    (2 bytes)
-            segmentLength           = obj.getNumericShort(startByte);
+            segmentLength           = double(obj.getNumericShort(startByte));
             % P     (1 byte)
             dataByteSize            = obj.inputStruct.numericData(startByte + 2);
             % Y     (2 byte)
@@ -403,7 +403,7 @@ classdef decoder < handle
             % luminance channels and one for chroma.
 
             % Lq
-            segmentLength                   = obj.getNumericShort(startByte);
+            segmentLength                   = double(obj.getNumericShort(startByte));
             if obj.verbose
                 disp(['- Quantisation Table Segment (' num2str(segmentLength) ' bytes)']);
             end
@@ -413,7 +413,7 @@ classdef decoder < handle
                 % Pq:Tq
                 [tablePrecision tableID]    = obj.getNumericNibblesFromByte(currentByte);
                 % Entries
-                obj.quantisationTables{tableID + 1} = TransformCoding.coefficientOrdering(obj.inputStruct.numericData(currentByte+1:currentByte+64), 'dezigzag');
+                obj.quantisationTables{tableID + 1} = double(TransformCoding.coefficientOrdering(obj.inputStruct.numericData(currentByte+1:currentByte+64), 'dezigzag'));
                 if obj.verbose
                     disp(['-- Table ' num2str(tableID) ': ']);
 
@@ -432,7 +432,7 @@ classdef decoder < handle
         function endByte = decodeHuffmanTablesFromNumericData(obj, startByte)
 
 
-            segmentLength = obj.getNumericShort(startByte);
+            segmentLength = double(obj.getNumericShort(startByte));
             if obj.verbose
                 disp(['- Huffman Table Segment (' num2str(segmentLength) ' bytes)']);
             end
@@ -443,7 +443,7 @@ classdef decoder < handle
                 [tableType tableID]    = obj.getNumericNibblesFromByte(currentByte);
 
                 % Li
-                Li = obj.inputStruct.numericData(currentByte + 1:currentByte + 16);
+                Li = double(obj.inputStruct.numericData(currentByte + 1:currentByte + 16));
 
                 currentByte = currentByte + 17;
 
@@ -503,7 +503,7 @@ classdef decoder < handle
             % NOTE: this only supports non-interleaved data at the moment
 
             % Ls    (2 bytes)
-            segmentLength = obj.getNumericShort(startByte);
+            segmentLength = double(obj.getNumericShort(startByte));
 
             if obj.verbose
                 disp(['- Scan Segment (header ' num2str(segmentLength) ' bytes):']);
@@ -550,12 +550,12 @@ classdef decoder < handle
             % Ref: CCITT Rec. T.81 (1992 E)	p.107
             % DC
             BITS = obj.huffmanDCCodeCountPerCodeLength{huffmanDCTableID + 1};
-            HUFFVALDC = cell2mat(obj.huffmanDCSymbolValuesPerCode{huffmanDCTableID + 1});
+            HUFFVALDC = double(cell2mat(obj.huffmanDCSymbolValuesPerCode{huffmanDCTableID + 1}));
             HUFFCODE = obj.huffcodeForDCCellArray{huffmanDCTableID + 1};
             [minCodeForDC maxCodeForDC valueTablePointerForDC] = EntropyCoding.generateDecodingProcedureTable(BITS, HUFFCODE);
 
             BITS = obj.huffmanACCodeCountPerCodeLength{huffmanACTableID + 1};
-            HUFFVALAC = cell2mat(obj.huffmanACSymbolValuesPerCode{huffmanACTableID + 1});
+            HUFFVALAC = double(cell2mat(obj.huffmanACSymbolValuesPerCode{huffmanACTableID + 1}));
             HUFFCODE = obj.huffcodeForACCellArray{huffmanACTableID + 1};
             [minCodeForAC maxCodeForAC valueTablePointerForAC] = EntropyCoding.generateDecodingProcedureTable(BITS, HUFFCODE);
 
@@ -570,11 +570,12 @@ classdef decoder < handle
             currentBit = 0;
 
             % compute total pixels for channel with its Hi,Vi and image w,h
-            totalBlocks = obj.componentSizeInBlocks{channelID};
+            totalBlocks = double(obj.componentSizeInBlocks{channelID});
 
             for i=1:totalBlocks(1)*totalBlocks(2)
                 % For given block
                 % DC Category Huffman decode
+
                 [categoryOfDCDiff currentByte currentBit] = EntropyCoding.decodeValue( obj.inputStruct.numericData, currentByte, currentBit, minCodeForDC, maxCodeForDC, valueTablePointerForDC, HUFFVALDC );
 
                 lengthOfExtraBits = categoryOfDCDiff;
